Adjust `for-each-package` argument positions
The positions of the arguments for the `for-each-package` kmono-build
API were chosen based on how cljfmt formats the usage of the macro.

I recently discovered that this formatting can be controlled (at least
for clojure-lsp) by setting the `:style/indent` metadata on the macro.

This allows for a nicer API with expected formatting. As this is a
breaking change I took the opportunity to adjust the `opts` map as well.