Add custom cmd parser built on org.babashka/cli
Currently kmono-cli is using cli-matic for parsing command line
args/opts. Internally cli-matic uses org.clojure/tools.cli which is
quite limited in it's flexibility.

It means we can't do things like parse both the below variants:

`kmono clojure -A`
`kmono clojure -A :nrepl`

We would need to do one or the other.

This change adds a new command spec parser which is built on top of
org.babashka/cli which gives us a lot more flexibility to parse opts in
a variety of ways.