Remove the requirement for `:kmono/package`
It's currently required for all packages that want to participate in the
workspace to contain a `:kmono/package` key with at minimum an empty
map.

This is a lot of additional unnecessary boilerplate.

This commit makes the key optional and therefore all packages found will
be automatically participating. To exclude a package from the workspace
you can use the new `:excluded` key:

```clojure
{:kmono/package {:excluded true}}
```