Improve performance of glob path matcher
This replaces the usage of `fs/glob` with a besboke implementation based
on the original.

Primarily this implementation allows combining multiple glob patterns
to avoid traversing the file tree multiple times.

This also directly matches for `deps.edn` files which avoids the need
for a second pass on the final results.