Traverse full graph when searching for cycles
This fixes a bug where some edges would be skipped for certain graph
configurations resulting in missed cycles.

This would eventually result in runtime Stack Overflow exceptions when
attempting to operate on the graph.

This also improves the error reporting in the case of a cycle.
Previously we would just report whatever was on the path when a cycle
was detected. Now we report only the exact cycle path.